+++ packages/manager/modules/pci/src/components/project/regions-list/regions-list.component.js
@@ -12,5 +12,9 @@ export default {
     displaySelectedRegion: '<',
     quotaUrl: '<?',
     currentRegion: '<?',
+    deploymentMode: '<?',
+    projectId: '<?',
+    reload: '<?',
+    loadEnd: '&?',
   },
 };