+++ packages/manager/apps/catalog/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.14.13](https://github.com/ovh/manager/compare/@ovh-ux/manager-catalog-app@1.14.12...@ovh-ux/manager-catalog-app@1.14.13) (2024-10-08)
+
+**Note:** Version bump only for package @ovh-ux/manager-catalog-app
+
+
+
+
+
 ## [1.14.12](https://github.com/ovh/manager/compare/@ovh-ux/manager-catalog-app@1.14.11...@ovh-ux/manager-catalog-app@1.14.12) (2024-10-04)
 
 **Note:** Version bump only for package @ovh-ux/manager-catalog-app