+++ packages/manager/apps/nutanix/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.45](https://github.com/ovh/manager/compare/@ovh-ux/manager-nutanix-app@2.2.44...@ovh-ux/manager-nutanix-app@2.2.45) (2024-10-08)
+
+**Note:** Version bump only for package @ovh-ux/manager-nutanix-app
+
+
+
+
+
 ## [2.2.44](https://github.com/ovh/manager/compare/@ovh-ux/manager-nutanix-app@2.2.43...@ovh-ux/manager-nutanix-app@2.2.44) (2024-10-03)
 
 **Note:** Version bump only for package @ovh-ux/manager-nutanix-app