+++ packages/manager/apps/pci-gateway/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.9.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-gateway-app@0.8.6...@ovh-ux/manager-pci-gateway-app@0.9.0) (2024-10-08)
+
+
+### Features
+
+* **pci-object-storage:** storage 3az ([1ed7551](https://github.com/ovh/manager/commit/1ed7551152ad155fd43484bca27217ae9cf80540))
+
+
+
+
+
 ## [0.8.6](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-gateway-app@0.8.5...@ovh-ux/manager-pci-gateway-app@0.8.6) (2024-10-04)
 
 **Note:** Version bump only for package @ovh-ux/manager-pci-gateway-app