+++ packages/manager/apps/vps/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.18](https://github.com/ovh/manager/compare/@ovh-ux/manager-vps-app@2.3.17...@ovh-ux/manager-vps-app@2.3.18) (2024-10-08)
+
+**Note:** Version bump only for package @ovh-ux/manager-vps-app
+
+
+
+
+
 ## [2.3.17](https://github.com/ovh/manager/compare/@ovh-ux/manager-vps-app@2.3.16...@ovh-ux/manager-vps-app@2.3.17) (2024-10-03)
 
 **Note:** Version bump only for package @ovh-ux/manager-vps-app